RT #55746 - remove bogus "... if %Config::Config" condition
This condition is always false for Perl 5.8.2 because of a specific bug in the
handling of tied hashes there. It causes M::SD's test suite to fail on Perl 5.8.2.
In all other cases this condition should be true (and if it isn't,
leaving the variable undefined isn't the correct way to handle this anyway).