Add package level caches for INC searches
The top level cache keys are used in the event
the search paths are changed mid-run.

They currently do not filter BSDPAN but
this is not needed since the join is
for key generation, not searching.

If a BSDPAN entry is added or removed then
a new top-level cache key will then be used.